3x speed boost

+++ include/config.h
@@ -7,10 +7,10 @@
 
 namespace CONFIG {
 	constexpr double INTAKE_RATIO = 66.0/96.0; // 66 chain links
-	constexpr double DRIVE_NOISE = 0.2;
+	constexpr float DRIVE_NOISE = 0.2;
 	constexpr Angle ANGLE_NOISE = 0.05_rad;
 	constexpr QLength DRIVE_RADIUS = 3.25_in/2.0;
-	constexpr double DRIVE_RATIO = 48.0/36.0;
+	constexpr float DRIVE_RATIO = 48.0/36.0;
 	constexpr double LIFT_RATIO = 8.0;
 	constexpr QLength TRACK_WIDTH = 10_in;
 	constexpr size_t NUM_PARTICLES = 1000;
@@ -27,11 +27,11 @@ namespace CONFIG {
 		return (velocity / 85_in/second).getValue() + (accel / (530_in/second/second)).getValue() + copysign(0.08, velocity.getValue());
 	}
 
-	const Eigen::Vector2d LINE_SENSOR_1_OFFSET((-1.2_in).getValue(), (-0.3_in).getValue());
+	const Eigen::Vector2f LINE_SENSOR_1_OFFSET((-1.2_in).getValue(), (-0.3_in).getValue());
 
-	const Eigen::Vector3d DISTANCE_LEFT_OFFSET((5.8_in).getValue(), (-4.2_in).getValue(), (90_deg).getValue());
-	const Eigen::Vector3d DISTANCE_BACK_OFFSET((-5.2_in).getValue(), (-5.4_in).getValue(), (180_deg).getValue());
-	const Eigen::Vector3d DISTANCE_RIGHT_OFFSET((-5.8_in).getValue(), (-4.2_in).getValue(), (-90_deg).getValue());
+	const Eigen::Vector3f DISTANCE_LEFT_OFFSET((5.8_in).getValue(), (-4.2_in).getValue(), (90_deg).getValue());
+	const Eigen::Vector3f DISTANCE_BACK_OFFSET((-5.2_in).getValue(), (-5.4_in).getValue(), (180_deg).getValue());
+	const Eigen::Vector3f DISTANCE_RIGHT_OFFSET((-5.8_in).getValue(), (-4.2_in).getValue(), (-90_deg).getValue());
 
-	const Eigen::Vector3d GPS_OFFSET((-6_in).getValue(), (-4_in).getValue(), (-90_deg).getValue());
+	const Eigen::Vector3f GPS_OFFSET((-6_in).getValue(), (-4_in).getValue(), (-90_deg).getValue());
 }
